chore: rename utils file
