+++ src/components/DataRoom/Hero/utils/AsciiEffect.ts
